refactor: 메인 뷰를 ChannelListView와 분할
